@@ -6,8 +6,8 @@
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "16.0.0-alpha.6",
-    "react-native": "0.44.0",
+    "react": "^16.4.0",
+    "react-native": "^0.55.4",
     "react-native-webrtc": "git+https://git@github.com/oney/react-native-webrtc.git#M63",
     "socket.io-client": "^1.3.7"
   }
