@@ -232,10 +232,13 @@ function getStats() {
 
 let container;
 
-const RCTWebRTCDemo = React.createClass({
-  getInitialState: function() {
+//const RCTWebRTCDemo = React.createClass({
+class RCTWebRTCDemo extends React.Component {
+  ds = null
+  constructor(props) {
+    super(props);
     this.ds = new ListView.DataSource({rowHasChanged: (r1, r2) => true});
-    return {
+    this.state = {
       info: 'Initializing',
       status: 'init',
       roomID: '',
@@ -246,15 +249,15 @@ const RCTWebRTCDemo = React.createClass({
       textRoomData: [],
       textRoomValue: '',
     };
-  },
-  componentDidMount: function() {
+  }
+  componentDidMount() {
     container = this;
-  },
+  }
   _press(event) {
-    this.refs.roomID.blur();
+    // this.refs.roomID.blur();
     this.setState({status: 'connect', info: 'Connecting'});
     join(this.state.roomID);
-  },
+  }
   _switchVideoType() {
     const isFront = !this.state.isFront;
     this.setState({isFront});
@@ -274,12 +277,12 @@ const RCTWebRTCDemo = React.createClass({
         pc && pc.addStream(localStream);
       }
     });
-  },
+  }
   receiveTextData(data) {
     const textRoomData = this.state.textRoomData.slice();
     textRoomData.push(data);
     this.setState({textRoomData, textRoomValue: ''});
-  },
+  }
   _textRoomPress() {
     if (!this.state.textRoomValue) {
       return
@@ -291,7 +294,7 @@ const RCTWebRTCDemo = React.createClass({
       pc.textDataChannel.send(this.state.textRoomValue);
     }
     this.setState({textRoomData, textRoomValue: ''});
-  },
+  }
   _renderTextRoom() {
     return (
       <View style={styles.listViewContainer}>
@@ -310,7 +313,7 @@ const RCTWebRTCDemo = React.createClass({
         </TouchableHighlight>
       </View>
     );
-  },
+  }
   render() {
     return (
       <View style={styles.container}>
@@ -338,7 +341,7 @@ const RCTWebRTCDemo = React.createClass({
               value={this.state.roomID}
             />
             <TouchableHighlight
-              onPress={this._press}>
+              onPress={this._press.bind(this)}>
               <Text>Enter room</Text>
             </TouchableHighlight>
           </View>) : null
@@ -352,7 +355,7 @@ const RCTWebRTCDemo = React.createClass({
       </View>
     );
   }
-});
+}
 
 const styles = StyleSheet.create({
   selfView: {
